chore: rename loadImage to load-image
